Remove unused functions from utils and clean up parse_attributes
